chore(changelog): note connector move
