+++ apps/web/app/connectors/bcorp/transform.test.ts
@@ -1,6 +1,6 @@
 // Use relative import to avoid relying on test-time path alias resolution.
 import { expect, test } from "vitest";
-import { transformBcorpRecord } from "../../../lib/connectors/bcorp/transform";
+import { transformBcorpRecord } from "../../../src/lib/connectors/bcorp/transform"; // adjust path after moving connector
 
 // Basic smoke test to ensure transform handles partial records without throwing.
 test("transform tolerates sparse input", () => {
