+++ apps/web/app/connectors/ecolabels/components/ResultCard.tsx
@@ -1,4 +1,4 @@
-import { CertifiedProduct } from "@/app/../src/lib/connectors/ecolabels/types";
+import { CertifiedProduct } from "@/src/lib/connectors/ecolabels/types"; // use alias to track shared types
 function Badge({ children }: { children: React.ReactNode }) {
   return <span className="inline-block text-xs px-2 py-1 rounded bg-gray-100 border">{children}</span>;
 }
