+++ apps/web/app/connectors/ecolabels/components/SearchBar.tsx
