+++ apps/web/app/connectors/fairtrade/components/ResultCard.tsx
