+++ apps/web/app/connectors/fairtrade/components/ResultTable.tsx
