+++ apps/web/app/connectors/fairtrade/components/SearchBar.tsx
