+++ apps/web/app/connectors/off/components/ResultCard.tsx
