+++ changelog/000-connectors-ui.md
@@ -1,0 +1,1 @@
+refactor: move connector pages to app directory and update imports (#000)
